neovim build files

+++ neovim-build.sh
@@ -1,7 +1,18 @@
 #!/bin/bash
-
 # exit when any command fails
 set -e
+
+if [[ -z $1 ]]
+then
+    echo Provide the path for install location!
+    exit(-1)
+fi
+
+if [[ $1 == "--help" ]] || [[ $1 == "-h" ]]
+then
+    echo "neovim-build.sh <install-path>"
+    exit(0)
+fi
 
 sudo rm /usr/bin/python
 sudo ln -s /usr/bin/python3 /usr/bin/python
@@ -15,6 +26,6 @@ trap finish EXIT INT TERM
 
 export PATH=/usr/bin:$PATH
 alias python='/usr/bin/python3'
-make -j`nproc` CMAKE_INSTALL_PREFIX=$HOME/neovim-dev
+make -j`nproc` CMAKE_INSTALL_PREFIX=$1 CMAKE_BUILD_TYPE=RelWithDebInfo
 make install
